requirements.txt
Flask>=2.0.0
flask-sqlalchemy>=2.5.1
flask-jwt-extended>=4.2.0
flask-cors>=3.0.10
openai>=0.27.0
scikit-learn>=1.0.0
nltk>=3.7
psycopg2-binary>=2.9.0
redis>=4.0.0
bcrypt>=3.2.0
python-dotenv>=0.19.0
gunicorn>=20.1.0
pytest>=6.2.0
pytest-cov>=2.12.0
requests>=2.26.0 